# packages/shared-contract/openapi/agentum.openapi.yaml
openapi: 3.1.0
info:
  title: Agentum API
  version: 0.1.0
  description: Agentum 平台 API 契约占位文件。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /api/tenants/{tenantId}/workflows/drafts:
    get:
      summary: 查询工作流草稿列表
      description: 流程设计页使用的租户内工作流定义草稿列表，后端按租户上下文和流程设计权限复核。
      tags:
        - Workflow
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          schema:
            type: string
            default: updatedAt,desc
      responses:
        "200":
          description: 工作流草稿分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WorkflowDraftPageResponse"
    post:
      summary: 新建工作流草稿
      tags:
        - Workflow
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowDraftRequest"
      responses:
        "200":
          description: 新建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WorkflowDraftRow"
  /api/tenants/{tenantId}/workflows/drafts/{workflowId}:
    get:
      summary: 获取工作流草稿详情
      tags:
        - Workflow
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 草稿详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WorkflowDraftDetail"
  /api/tenants/{tenantId}/workflows/drafts/{workflowId}/graph:
    put:
      summary: 保存工作流草稿节点与边
      tags:
        - Workflow
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWorkflowDraftGraphRequest"
      responses:
        "200":
          description: 保存成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WorkflowDraftDetail"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: WORKFLOW_NODE_CONFIG_INVALID
        message:
          type: string
          example: 智能体节点缺少输出变量配置
        details:
          type: object
      required:
        - code
        - message
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          anyOf:
            - $ref: "#/components/schemas/ApiError"
            - type: "null"
        requestId:
          type: string
      required:
        - success
        - data
        - error
        - requestId
...
    WorkflowDraftPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowDraftRow"
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
          format: int64
        totalPages:
          type: integer
      required:
        - items
        - page
        - size
        - total
        - totalPages
    WorkflowDraftRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, review, published]
        nodeCount:
          type: integer
        pausePointCount:
          type: integer
        ownerName:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - description
        - status
        - nodeCount
        - pausePointCount
        - ownerName
        - updatedAt
    CreateWorkflowDraftRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 1000
      required:
        - name
    WorkflowDraftDetail:
      type: object
      properties:
        draft:
          $ref: "#/components/schemas/WorkflowDraftRow"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNodeDraft"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdgeDraft"
      required:
        - draft
        - nodes
        - edges
    SaveWorkflowDraftGraphRequest:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNodeDraft"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdgeDraft"
      required:
        - nodes
        - edges
    WorkflowNodeDraft:
      type: object
      properties:
        nodeId:
          type: string
        nodeType:
          type: string
          enum: [trigger, user_input, agent, parallel_group, merge, condition, human_review, delivery]
        name:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        inputVariables:
          type: array
          items:
            type: string
        outputVariables:
          type: array
          items:
            type: string
        config:
          type: object
      required:
        - nodeId
        - nodeType
        - name
        - positionX
        - positionY
        - inputVariables
        - outputVariables
        - config
    WorkflowEdgeDraft:
      type: object
      properties:
        edgeId:
          type: string
        sourceNodeId:
          type: string
        targetNodeId:
          type: string
        label:
          type: string
        conditionExpression:
          type: string
      required:
        - edgeId
        - sourceNodeId
        - targetNodeId
...